+++ libraries/rib-base/src/main/java/com/badoo/ribs/minimal/reactive/Cancellable.kt
@@ -1,0 +1,18 @@
+package com.badoo.ribs.minimal.reactive
+
+interface Cancellable {
+    fun cancel()
+
+    companion object {
+        fun cancellableOf(f: () -> Unit): Cancellable =
+            object : Cancellable {
+                override fun cancel() {
+                    f()
+                }
+            }
+    }
+
+    object Empty : Cancellable {
+        override fun cancel() { }
+    }
+}
