+++ libraries/rib-base/src/main/java/com/badoo/ribs/minimal/reactive/CompositeCancellable.kt
@@ -1,0 +1,13 @@
+package com.badoo.ribs.minimal.reactive
+
+class CompositeCancellable(vararg cancellables: Cancellable) : Cancellable {
+    private val items = mutableListOf(*cancellables)
+    operator fun plusAssign(item: Cancellable) {
+        items += item
+    }
+
+    override fun cancel() {
+        items.forEach { it.cancel() }
+        items.clear()
+    }
+}
