+++ libraries/rib-base/src/main/java/com/badoo/ribs/minimal/reactive/Emitter.kt
@@ -1,0 +1,5 @@
+package com.badoo.ribs.minimal.reactive
+
+interface Emitter<in T> {
+    fun emit(value: T)
+}
