+++ libraries/rib-base/src/main/java/com/badoo/ribs/minimal/reactive/Relay.kt
@@ -2,43 +2,6 @@ package com.badoo.ribs.minimal.reactive
 
 import com.badoo.ribs.minimal.reactive.Cancellable.Companion.cancellableOf
 
-interface Source<out T> {
-    fun observe(callback: (T) -> Unit): Cancellable
-}
-
-interface Emitter<in T> {
-    fun emit(value: T)
-}
-
-interface Cancellable {
-    fun cancel()
-
-    companion object {
-        fun cancellableOf(f: () -> Unit): Cancellable =
-            object : Cancellable {
-                override fun cancel() {
-                    f()
-                }
-            }
-    }
-
-    object Empty : Cancellable {
-        override fun cancel() { }
-    }
-}
-
-class CompositeCancellable(vararg cancellables: Cancellable) : Cancellable {
-    private val items = mutableListOf(*cancellables)
-    operator fun plusAssign(item: Cancellable) {
-        items += item
-    }
-
-    override fun cancel() {
-        items.forEach { it.cancel() }
-        items.clear()
-    }
-}
-
 
 class Relay<T> : Source<T>, Emitter<T> {
     private val listeners: MutableList<(T) -> Unit> = mutableListOf()
