+++ libraries/rib-base/src/main/java/com/badoo/ribs/minimal/reactive/Source.kt
@@ -1,0 +1,5 @@
+package com.badoo.ribs.minimal.reactive
+
+interface Source<out T> {
+    fun observe(callback: (T) -> Unit): Cancellable
+}
